backend/requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
web3==6.11.3
requests==2.31.0
pydantic==1.10.13
python-multipart==0.0.6
aiohttp==3.9.1

# Blockchain Interaction
eth-abi==5.0.0
eth-account==0.9.0
eth-hash==0.5.2
eth-typing==3.5.0
eth-utils==2.2.0 